Generate secret keyfile
Instead of using the same secret keyfile everywhere, this commit
uses the lookup password plugin in Ansible to generate a random
key for MongoDB.

I'm not sure if there's a better way than using `set_fact` to do this,
but it seems to work just fine. This could also be improved to use
vaults instead, but this method was the least complex.